chore(lint): migrate to oxlint
