+++ cloud-agent/package.json
@@ -14,7 +14,9 @@
     "prestart": "pnpm run build:wrapper",
     "start": "wrangler dev",
     "types": "wrangler types",
-    "lint": "eslint --config eslint.config.mjs --cache 'src/**/*.ts' 'wrapper/src/**/*.ts'",
+    "lint": "pnpm run lint:oxlint && pnpm run lint:eslint:fallback",
+    "lint:oxlint": "pnpm -w exec oxlint --config .oxlintrc.json cloud-agent/src cloud-agent/wrapper/src",
+    "lint:eslint:fallback": "eslint --config eslint.config.mjs --cache 'src/**/*.ts' 'wrapper/src/**/*.ts'",
     "lint:fix": "eslint --config eslint.config.mjs --cache --fix 'src/**/*.ts' 'wrapper/src/**/*.ts'",
     "format": "oxfmt src",
     "format:check": "oxfmt --list-different src",
