+++ cloudflare-app-builder/package.json
@@ -8,7 +8,9 @@
     "deploy": "wrangler deploy",
     "cf-typegen": "wrangler types --env-interface CloudflareEnv worker-configuration.d.ts",
     "typecheck": "tsgo --noEmit --incremental false",
-    "lint": "eslint --config eslint.config.mjs --cache 'src/**/*.ts'",
+    "lint": "pnpm run lint:oxlint && pnpm run lint:eslint:fallback",
+    "lint:oxlint": "pnpm -w exec oxlint --config .oxlintrc.json cloudflare-app-builder/src",
+    "lint:eslint:fallback": "eslint --config eslint.config.mjs --cache 'src/**/*.ts'",
     "test": "vitest run",
     "test:git": "tsx src/_integration_tests/test-git-integration.ts",
     "test:git-head": "tsx src/_integration_tests/test-git-head-and-history.ts",
