+++ cloudflare-gmail-push/package.json
@@ -6,7 +6,9 @@
     "deploy": "wrangler deploy",
     "test": "vitest run",
     "typecheck": "tsgo --noEmit --incremental false",
-    "lint": "eslint src/",
+    "lint": "pnpm run lint:oxlint && pnpm run lint:eslint:fallback",
+    "lint:oxlint": "pnpm -w exec oxlint --config .oxlintrc.json cloudflare-gmail-push/src",
+    "lint:eslint:fallback": "eslint src/",
     "cf-typegen": "wrangler types --env-file /dev/null"
   },
   "dependencies": {
