+++ cloudflare-o11y/package.json
@@ -9,7 +9,9 @@
     "test": "vitest",
     "cf-typegen": "wrangler types",
     "typecheck": "tsgo --noEmit --incremental false",
-    "lint": "eslint --config eslint.config.mjs --cache 'src/**/*.ts'"
+    "lint": "pnpm run lint:oxlint && pnpm run lint:eslint:fallback",
+    "lint:oxlint": "pnpm -w exec oxlint --config .oxlintrc.json cloudflare-o11y/src",
+    "lint:eslint:fallback": "eslint --config eslint.config.mjs --cache 'src/**/*.ts'"
   },
   "devDependencies": {
     "@cloudflare/vitest-pool-workers": "^0.12.21",
