+++ cloudflare-security-auto-analysis/package.json
@@ -6,7 +6,9 @@
     "deploy:prod": "wrangler deploy",
     "deploy:dev": "wrangler deploy --env dev",
     "dev": "wrangler dev --env dev",
-    "lint": "eslint --config eslint.config.mjs --cache 'src/**/*.ts'",
+    "lint": "pnpm run lint:oxlint && pnpm run lint:eslint:fallback",
+    "lint:oxlint": "pnpm -w exec oxlint --config .oxlintrc.json cloudflare-security-auto-analysis/src",
+    "lint:eslint:fallback": "eslint --config eslint.config.mjs --cache 'src/**/*.ts'",
     "typecheck": "tsgo --noEmit --incremental false",
     "types": "wrangler types --env-interface CloudflareEnv worker-configuration.d.ts",
     "test": "vitest run"
