+++ cloudflare-webhook-agent-ingest/package.json
@@ -11,7 +11,9 @@
     "dev": "wrangler dev --env dev",
     "start": "wrangler dev --env dev",
     "types": "wrangler types",
-    "lint": "eslint --config eslint.config.mjs --cache 'src/**/*.ts'",
+    "lint": "pnpm run lint:oxlint && pnpm run lint:eslint:fallback",
+    "lint:oxlint": "pnpm -w exec oxlint --config .oxlintrc.json cloudflare-webhook-agent-ingest/src",
+    "lint:eslint:fallback": "eslint --config eslint.config.mjs --cache 'src/**/*.ts'",
     "lint:fix": "eslint --config eslint.config.mjs --cache --fix 'src/**/*.ts'",
     "format": "oxfmt src",
     "format:check": "oxfmt --list-different src",
