+++ kiloclaw/package.json
@@ -12,7 +12,9 @@
     "dev": "wrangler dev",
     "types": "wrangler types",
     "typecheck": "tsgo --noEmit --incremental false",
-    "lint": "eslint --config eslint.config.mjs --cache 'src/**/*.ts'",
+    "lint": "pnpm run lint:oxlint && pnpm run lint:eslint:fallback",
+    "lint:oxlint": "pnpm -w exec oxlint --config .oxlintrc.json kiloclaw/src",
+    "lint:eslint:fallback": "eslint --config eslint.config.mjs --cache 'src/**/*.ts'",
     "lint:fix": "eslint --config eslint.config.mjs --cache --fix 'src/**/*.ts'",
     "format": "oxfmt src",
     "format:check": "oxfmt --list-different src",
