+++ package.json
@@ -11,7 +11,9 @@
     "build": "next build",
     "start": "next start",
     "stripe": "stripe listen --forward-to http://localhost:3000/api/stripe/webhook",
-    "lint": "NODE_OPTIONS='--max-old-space-size=8192' eslint --config eslint.config.mjs --cache 'src/**/*.{ts,tsx,js,jsx}' && pnpm -r --filter '!kilocode-backend' --filter '!storybook' run lint",
+    "lint": "pnpm run lint:oxlint && pnpm run lint:eslint:fallback && pnpm -r --filter '!kilocode-backend' --filter '!storybook' run lint",
+    "lint:oxlint": "oxlint --config .oxlintrc.json src",
+    "lint:eslint:fallback": "NODE_OPTIONS='--max-old-space-size=8192' eslint --config eslint.fallback.config.mjs --cache --cache-strategy content 'src/**/*.{ts,tsx}'",
     "format": "oxfmt .",
     "format:check": "oxfmt --list-different .",
     "format:changed": "git diff --name-only $(git merge-base origin/main HEAD) --diff-filter=ACMR -- '**/*.js' '**/*.jsx' '**/*.ts' '**/*.tsx' '**/*.json' '**/*.css' '**/*.md' | xargs -r oxfmt --no-error-on-unmatched-pattern",
@@ -169,15 +171,17 @@
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-drizzle": "^0.2.3",
     "eslint-plugin-n": "^17.24.0",
+    "eslint-plugin-oxlint": "^1.55.0",
     "eslint-plugin-react-hooks": "^7.0.1",
     "eslint-plugin-tailwindcss": "4.0.0-beta.0",
     "husky": "^9.1.7",
     "jest": "^30.3.0",
     "knip": "^5.86.0",
     "madge": "^8.0.0",
+    "oxfmt": "^0.40.0",
+    "oxlint": "^1.55.0",
     "p-limit": "^7.3.0",
     "postcss": "^8.5.8",
-    "oxfmt": "^0.40.0",
     "tailwindcss": "^4.2.1",
     "ts-jest": "^29.4.6",
     "tsconfig-paths": "^4.2.0",
